requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
rapidfuzz==3.5.2
flask==3.0.0
python-dotenv==1.0.0 